Extract DocumentSpanExtensions to its own file
